feat: replace keypair and ursa-optional with web-crypto for node
Removes native module `ursa-optional` and js fallback library `keypair`
in favour of using web-crypto which is available in node 15+

Also adds benchmark suite similar to Ed25519:

```console
$ node index.js
ursa-optional x 323 ops/sec ±2.84% (70 runs sampled)
keypair x 55.33 ops/sec ±11.75% (59 runs sampled)
node.js web-crypto x 300 ops/sec ±2.29% (74 runs sampled)
fastest is ursa-optional
```

BREAKING CHANGE: requires node 15+
