+++ package.json
@@ -10,7 +10,7 @@
     "./src/ciphers/aes-gcm.js": "./src/ciphers/aes-gcm.browser.js",
     "./src/hmac/index.js": "./src/hmac/index-browser.js",
     "./src/keys/ecdh.js": "./src/keys/ecdh-browser.js",
-    "./src/keys/rsa.js": "./src/keys/rsa-browser.js"
+    "./src/webcrypto.js": "./src/webcrypto-browser.js"
   },
   "files": [
     "src",
@@ -44,13 +44,11 @@
     "@noble/secp256k1": "^1.3.0",
     "err-code": "^3.0.1",
     "iso-random-stream": "^2.0.0",
-    "keypair": "^1.0.4",
     "multiformats": "^9.4.5",
     "node-forge": "^0.10.0",
     "pem-jwk": "^2.0.0",
     "protobufjs": "^6.11.2",
-    "uint8arrays": "^3.0.0",
-    "ursa-optional": "^0.10.1"
+    "uint8arrays": "^3.0.0"
   },
   "devDependencies": {
     "@types/mocha": "^9.0.0",
