--- test/node.js
@@ -1,3 +1,0 @@
-'use strict'
-
-require('./keys/rsa-crypto-libs')
